# openapi.yml
openapi: 3.0.0
info:
  title: Pawn Manager API
  description: API for managing pawn shop operations, including loans, customers, payments, collateral, disbursements, and stores.
  version: '1.0'
servers:
  - url: /v1
    description: V1 API Endpoint
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        totalItems:
          type: integer
        totalPages:
          type: integer
        currentPage:
          type: integer
        itemsPerPage:
          type: integer
...
    StoreResponseWrapper:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/StoreResponse'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaymentMethod:
      type: string
      enum: [CASH, BANK_TRANSFER]
    PaymentType:
      type: string
      enum: [PERIODIC, EARLY, PAYOFF, ADJUSTMENT]
    PaymentRequestDto:
      type: object
      required:
        - loanId
        - amount
        - paymentMethod
        - paymentType
      properties:
        loanId:
          type: string
          format: uuid
        amount:
          type: number
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
        paymentType:
          $ref: '#/components/schemas/PaymentType'
        notes:
          type: string
    PaymentResponse:
      type: object
      description: Simplified response
      properties:
        id:
          type: string
        amount:
          type: number
        status:
          type: string
    PaymentListItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        loanId:
          type: string
          format: uuid
        storeId:
          type: string
          format: uuid
        amount:
          type: number
        paymentType:
          $ref: '#/components/schemas/PaymentType'
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
        paidAt:
          type: string
          format: date-time
        referenceCode:
          type: string
    ListPaymentsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PaymentListItem'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    # --- DISBURSEMENT ---
    DisbursementMethod:
      type: string
      enum: [CASH, BANK_TRANSFER]
    DisbursementRequestDto:
      type: object
      required:
        - loanId
        - storeId
        - amount
        - disbursementMethod
        - recipientName
      properties:
        loanId:
          type: string
          format: uuid
        storeId:
          type: string
          format: uuid
        amount:
          type: number
        disbursementMethod:
          $ref: '#/components/schemas/DisbursementMethod'
        recipientName:
          type: string
        recipientIdNumber:
          type: string
        witnessName:
          type: string
        bankTransferRef:
          type: string
        bankAccountNumber:
          type: string
        bankName:
          type: string
        notes:
          type: string
    DisbursementListItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        loanId:
          type: string
          format: uuid
        amount:
          type: number
        disbursementMethod:
          $ref: '#/components/schemas/DisbursementMethod'
        disbursedAt:
          type: string
        referenceCode:
          type: string
        recipientName:
          type: string
        loanCode:
          type: string
    DisbursementDetailsResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        loanId:
          type: string
        loanCode:
          type: string
        amount:
          type: number
        disbursementMethod:
          $ref: '#/components/schemas/DisbursementMethod'
        disbursedAt:
          type: string
        referenceCode:
          type: string
        recipientName:
          type: string
        recipientIdNumber:
          type: string
        bankTransferRef:
          type: string
        bankAccountNumber:
          type: string
        bankName:
          type: string
        notes:
          type: string
    ListDisbursementsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/DisbursementListItem'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    DisbursementResponseWrapper:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/DisbursementDetailsResponse'

    # --- COLLATERAL ---
...
    CreateStoreDTO:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        address:
          type: string
        storeInfo:
          type: object
        isActive:
          type: boolean
          default: true
    UpdateStoreDTO:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        isActive:
          type: boolean
    StoreResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: string
        isActive:
          type: boolean
    StoreListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/StoreResponse'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    # --- REPORTS ---
...
    ProvinceResponse:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
    WardResponse:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        parentId:
          type: string

    # --- CONFIGURATIONS ---
...
security:
  - bearerAuth: []
paths:
...
  /payments:
    get:
      summary: List payments
      tags: [Payments]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
        - name: loanId
          in: query
          schema:
            type: string
        - name: storeId
          in: query
          schema:
            type: string
        - name: paymentMethod
          in: query
          schema:
            $ref: '#/components/schemas/PaymentMethod'
        - name: paymentType
          in: query
          schema:
            $ref: '#/components/schemas/PaymentType'
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
        - name: minAmount
          in: query
          schema:
            type: number
        - name: maxAmount
          in: query
          schema:
            type: number
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListPaymentsResponse'
    post:
      summary: Create Payment
      tags: [Payments]
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequestDto'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '409':
          description: Duplicate Idempotency Key

  # --- DISBURSEMENT ---
  /disbursements:
    get:
      summary: List disbursements
      tags: [Disbursements]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: loanId
          in: query
          schema:
            type: string
            format: uuid
        - name: storeId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListDisbursementsResponse'
    post:
      summary: Create disbursement
      tags: [Disbursements]
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisbursementRequestDto'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisbursementResponseWrapper'

  # --- COLLATERAL ---
...
  /stores:
    get:
      summary: List stores
      tags: [Stores]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreListResponse'
    post:
      summary: Create store
      tags: [Stores]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStoreDTO'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreResponseWrapper'
  /stores/{id}:
    get:
      summary: Get store
      tags: [Stores]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreResponseWrapper'
    patch:
      summary: Update store
      tags: [Stores]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStoreDTO'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreResponseWrapper'

  # --- REPORTS ---
...
  /provinces:
    get:
      summary: Get provinces
      tags: [Location]
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProvinceResponse'
  /provinces/{code}/wards:
    get:
      summary: Get wards by province code
      tags: [Location]
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WardResponse'

  # --- CONFIGURATIONS ---
...